Moved the code that retrieves / calculates the AssessmentScore
